% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupvar_to_model}
\alias{groupvar_to_model}
\title{Mapping of groupvars to model}
\usage{
groupvar_to_model(pred, groupvars)
}
\arguments{
\item{pred}{data.frame with at least one column called model}

\item{groupvars}{quoted variable names used to define groups within
which models should be ranked.}
}
\value{
list
}
\description{
Mapping of groupvar to models run
}
\details{
In some cases, we might not have the same set of models run for
a grouping variable. For example, say the grouping variable is
country, and we have Models \eqn{m_1}, \eqn{m_2} and \eqn{m_3}
run for country \eqn{c_1}, and models \eqn{m_1, m_2, m_3, m_4} run
country \eqn{c_2}. For this example, this function will return the
list
{c_1 = {m_1, m_2, m_3}, c_2 = {m_1, m_2, m_3, m_4}}
This can be used to identify which countries have the same set of
models run for them.
}
\author{
Sangeeta Bhatia
}
