% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_weights.R
\name{model_ranks}
\alias{model_ranks}
\title{Assign ranks to models}
\usage{
model_ranks(pred, groupvars, errvar)
}
\arguments{
\item{pred}{data.frame with at least one column called model}

\item{groupvars}{quoted variable names used to define groups within
which models should be ranked.}

\item{errvar}{quoted column name that contains the metric used to
rank models}
}
\value{
data.frame with the same structure as \code{pred} and an
extra column called rank which contains the model ranks within
the groups defined by \code{groupvar}
}
\description{
Rank models according to the error
}
\details{
When multiple models have made predictions, we want to rank them
by their error, where error can be any metric e.g. Root mean
squared error or likelihood. Each model is ranked within the
groups defined by \code{groupvars}
}
\examples{
pred <- data.frame(
  models = as.factor(c("a", "b", "c", "a", "b", "c")),
  country = c("C1", "C1", "C1", "C2", "C2", "C2"),
  error = c(6L, 1L, 3L, 4L, 7L, 9L),
  date = c("1", "1", "1", "2", "2", "2"),
  stringsAsFactors = FALSE
)
model_ranks(pred, c("country", "date"), "error")
}
\author{
Sangeeta Bhatia
}
