% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_to_groupvar}
\alias{model_to_groupvar}
\title{Mapping of models to groups for which they have been run}
\usage{
model_to_groupvar(pred, groupvars)
}
\arguments{
\item{pred}{data.frame with at least one column called model}

\item{groupvars}{quoted variable names used to define groups within
which models should be ranked.}
}
\value{
list
}
\description{
Mapping of models to groups for which they have been run
}
\details{
In some cases, we might not have the same set of models run for
a grouping variable. For example, say the grouping variable is
country, and we have Models \eqn{M_1}, \eqn{M_2} and \eqn{M_3}
run for country \eqn{C_1}, and models \eqn{M_1, M_2, M_3, M_4} run
country \eqn{C_2}. For this example, this function will return the
list
{m_1 = {c_1, c_2}, m_2 = {c_1, c_1}, m_3 = {c_1, c_2}, m_4 = {c_2}}
This can be used to identify for which countries a give model has
been run.
}
\author{
Sangeeta Bhatia
}
