% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_weights.R
\name{model_weights_in_group}
\alias{model_weights_in_group}
\title{Compute Model Weights}
\usage{
model_weights_in_group(pred, groupvars = "model", rankvar = "rank")
}
\arguments{
\item{pred}{data.frame that has at least columns specified via
parameters \code{groupvars} and \code{rank}. pred is expected
to have the same number of elements of grouping variable within
each group. That is, each model should have been run for the same
number of elements of groupvar. If this is not the case, use
\code{model_weights} instead.}

\item{groupvars}{quoted variable names used to define groups within
which model weights are computed. Defaults to model.}

\item{rankvar}{quoted variable names that contain model ranks.}
}
\value{
a data.frame with columns model and weight where weight is
the unnormalised weight of the model
}
\description{
Compute model weights from model ranks
}
\details{
In a group defined by \code{groupvars}, the weight of a model is
defined as
\deqn{weight(M) = \sum\limits_{i = 1}^{M}{K(i) / i}}
where \eqn{K(i)} is the number of times model M is ranked \eqn{i}
among \eqn{M_1, M_2, \dots M} models.
}
\examples{

pred <- data.frame(
  model = as.factor(c("a", "b", "c", "a", "b", "c")),
  country = c("C1", "C1", "C1", "C2", "C2", "C2"),
  error = c(6L, 1L, 3L, 4L, 7L, 9L),
  date = c("1", "1", "1", "2", "2", "2"),
  stringsAsFactors = FALSE
)
ranked <- model_ranks(pred, c("country", "date"), "error")
model_weights_in_group(ranked)
}
\seealso{
\code{model_ranks} \code{model_weights}
}
\author{
Sangeeta Bhatia
}
