% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupvar_to_model}
\alias{groupvar_to_model}
\title{Mapping of groupvars to model}
\usage{
groupvar_to_model(pred, groupvars)
}
\arguments{
\item{pred}{data.frame with at least one column called model}

\item{groupvars}{quoted variable names used to define groups within
which models should be ranked.}
}
\value{
list
}
\description{
Mapping of groupvar to models run
}
\details{
In some cases, we might not have the same set of models run for
a grouping variable. For example, say the grouping variable is
country, and we have Models \eqn{m1}, \eqn{m2} and \eqn{m3}
run for country \eqn{c1}, and models \eqn{m1, m2, m3, m4} run
for country \eqn{c2}. For this example, \code{groupvar_to_model} returns
a named list where names are formed by concatenating models
and elements are countries for which a particular set of models has been run.
(m1_m2_m3 = c1, m1_m2_m3_m4 = c2)
}
\author{
Sangeeta Bhatia
}
