% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_weights.R
\name{model_weights}
\alias{model_weights}
\title{Compute Model Weights}
\usage{
model_weights(pred, groupvars, errvar)
}
\arguments{
\item{pred}{data.frame with at least one column called model}

\item{groupvars}{quoted variable names used to define groups within
which models should be ranked.}

\item{errvar}{quoted column name that contains the metric used to
rank models}
}
\value{
list
}
\description{
Compute Model Weights
}
\details{
Compute model weights within each group
}
\author{
Sangeeta Bhatia
}
